feat: Implement RFC-2758 Merge Append Into Write (#2880)
* Save work


* Fix test


* fix binding


* allow append for azblob


* Fix test


* Fix write test


* Make sure buffer has been flushed


* disable hdfs for now


* ftp doesn't support append yet


---------

+++ core/src/services/azblob/core.rs
@@ -365,7 +365,7 @@ impl AzblobCore {
     pub fn azblob_append_blob_request(
         &self,
         path: &str,
-        size: usize,
+        size: u64,
         position: Option<u64>,
         body: AsyncBody,
     ) -> Result<Request<AsyncBody>> {
